+++ src/main/java/com/okta/developer/jugtours/web/TableController.java
@@ -1,10 +1,12 @@
 package com.okta.developer.jugtours.web;
 
+import com.okta.developer.jugtours.enums.TableType;
+import com.okta.developer.jugtours.exception.ValidationCustomException;
 import com.okta.developer.jugtours.model.TableEntity;
+import com.okta.developer.jugtours.repository.BookEventRepository;
 import com.okta.developer.jugtours.repository.RestaurantRepository;
 import com.okta.developer.jugtours.repository.TableRepository;
 import jakarta.validation.Valid;
-import jakarta.websocket.server.PathParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
@@ -31,11 +33,14 @@ public class TableController {
   private final Logger log = LoggerFactory.getLogger(TableController.class);
   private final TableRepository tableRepository;
   private final RestaurantRepository restaurantRepository;
+  private final BookEventRepository bookEventRepository;
 
   public TableController(TableRepository tableRepository,
-    RestaurantRepository restaurantRepository) {
+    RestaurantRepository restaurantRepository,
+    BookEventRepository bookEventRepository) {
     this.tableRepository = tableRepository;
     this.restaurantRepository = restaurantRepository;
+    this.bookEventRepository = bookEventRepository;
   }
 
   @GetMapping
@@ -58,6 +63,21 @@ public class TableController {
 
     var restaurant = restaurantRepository.findById(restaurantId).get();
 
+    if (restaurant.getTables().size() >= TableType.values().length) {
+      throw new ValidationCustomException(
+        String.format("Max number of table types reached: %d. Posible table types: [%s]",
+          TableType.values().length, List.of(TableType.values())));
+    }
+
+    final var hasTheSameTableType = restaurant.getTables().stream()
+      .anyMatch(tableEntity -> tableEntity.getTableType().equals(table.getTableType()));
+    if (hasTheSameTableType) {
+      throw new ValidationCustomException(
+        String.format("This restaurant '%s' have already had table with the same table type '%s' "
+                      + "Please update existing or remove and create new one.",
+          restaurant.getName(), table.getTableType()));
+    }
+
     table.setRestaurant(restaurant);
     TableEntity result = tableRepository.save(table);
     return ResponseEntity.created(
@@ -65,7 +85,8 @@ public class TableController {
   }
 
   @PutMapping
-  ResponseEntity<TableEntity> updateTable(@Valid @RequestBody TableEntity table) {
+  ResponseEntity<TableEntity> updateTable(@Valid @RequestBody TableEntity table,
+    @RequestParam("restaurantId") Long restaurantId) {
     log.info("Request to update table: {}", table);
     TableEntity result = tableRepository.save(table);
     return ResponseEntity.ok().body(result);
@@ -74,6 +95,14 @@ public class TableController {
   @DeleteMapping("/{id}")
   public ResponseEntity<?> deleteTable(@PathVariable Long id) {
     log.info("Request to delete table: {}", id);
+
+    final var table = tableRepository.findById(id).get();
+    final var events = bookEventRepository.findAllByTableId(id);
+    if (!events.isEmpty()) {
+      throw new ValidationCustomException(String.format(
+        "Table '%s' can't be deleted. This table used in bookings.", table.getTableType().name()));
+    }
+
     tableRepository.deleteById(id);
     return ResponseEntity.ok().build();
   }
